:rocket: Backlight Support

+++ curves/DecreaseEaseOut.go
@@ -1,8 +1,8 @@
 package curves
 
-func DecreaseEaseOut(current uint8) uint8 {
+func DecreaseEaseOut(current uint16) uint16 {
 
-	var result uint8 = 100;
+	var result uint16 = 100;
 
 	if current >= 0 && current <= 1 {
 		result = 1;
