+++ curves/IncreaseEaseIn.go
@@ -1,8 +1,8 @@
 package curves
 
-func IncreaseEaseIn(current uint8) uint8 {
+func IncreaseEaseIn(current uint16) uint16 {
 
-	var result uint8 = 100;
+	var result uint16 = 100;
 
 	if current == 0 {
 		result = 1;
